Feat: reorganized project structure to match for requests in sub issues
